+++ backend/controllers/module.js
@@ -1,6 +1,7 @@
 const Module = require("../models/module");
 const State = require("../models/state");
-
+const jwt = require("jsonwebtoken");
+const config = require("../utils/config");
 exports.getModuleByModuleId = async (moduleId) => {
   return await Module.findById(moduleId).populate("questions");
 };
@@ -104,6 +105,10 @@ exports.getModuleByLicenseIdAPI = async (req, res, next) => {
 
 exports.createModuleAPI = async (req, res, next) => {
   try {
+    const token = req.token;
+    const decodeToken = jwt.verify(token, config.SECRET);
+    if(!decodeToken.id || !decodeToken.role) return res.status(403).json({message: "Token missing or invalid"});
+    if(decodeToken.role !== "ADMIN") return res.status(403).json({message: "Role is not allowed"});
     const module = req.body;
     const check = await Module.findOne({ name: module.name });
     if (check) {
@@ -118,6 +123,10 @@ exports.createModuleAPI = async (req, res, next) => {
 
 exports.deleteModuleAPI = async (req, res, next) => {
   try {
+    const token = req.token;
+    const decodeToken = jwt.verify(token, config.SECRET);
+    if(!decodeToken.id || !decodeToken.role) return res.status(403).json({message: "Token missing or invalid"});
+    if(decodeToken.role !== "ADMIN") return res.status(403).json({message: "Role is not allowed"});
     await Module.findByIdAndDelete(req.params.moduleId);
     res.status(200).json({ message: "Module deleted" });
   } catch (error) {
@@ -127,6 +136,10 @@ exports.deleteModuleAPI = async (req, res, next) => {
 
 exports.addModuleToStateAPI = async (req, res, next) => {
   try {
+    const token = req.token;
+    const decodeToken = jwt.verify(token, config.SECRET);
+    if(!decodeToken.id || !decodeToken.role) return res.status(403).json({message: "Token missing or invalid"});
+    if(decodeToken.role !== "ADMIN") return res.status(403).json({message: "Role is not allowed"});
     const { stateId, moduleId } = req.params;
     const module = await Module.findById(moduleId);
     if (!module) {
@@ -150,6 +163,10 @@ exports.addModuleToStateAPI = async (req, res, next) => {
 
 exports.removeModuleOfStateAPI = async (req, res, next) => {
   try {
+    const token = req.token;
+    const decodeToken = jwt.verify(token, config.SECRET);
+    if(!decodeToken.id || !decodeToken.role) return res.status(403).json({message: "Token missing or invalid"});
+    if(decodeToken.role !== "ADMIN") return res.status(403).json({message: "Role is not allowed"});
     const { stateId, moduleId } = req.params;
     const module = await Module.findOne({ _id: moduleId, states: stateId });
     if (!module) {
@@ -184,6 +201,10 @@ exports.getDetailModuleByModuleIdAPI = async (req, res, next) => {
 
 exports.updateModuleAPI = async (req, res, next) => {
   try {
+    const token = req.token;
+    const decodeToken = jwt.verify(token, config.SECRET);
+    if(!decodeToken.id || !decodeToken.role) return res.status(403).json({message: "Token missing or invalid"});
+    if(decodeToken.role !== "ADMIN") return res.status(403).json({message: "Role is not allowed"});
     const { moduleId } = req.params;
     const moduleUpdated = await Module.findByIdAndUpdate(moduleId, req.body);
     res.status(200).json(moduleUpdated);
@@ -194,6 +215,10 @@ exports.updateModuleAPI = async (req, res, next) => {
 
 exports.updatePositionModuleAPI = async (req, res, next) => {
   try {
+    const token = req.token;
+    const decodeToken = jwt.verify(token, config.SECRET);
+    if(!decodeToken.id || !decodeToken.role) return res.status(403).json({message: "Token missing or invalid"});
+    if(decodeToken.role !== "ADMIN") return res.status(403).json({message: "Role is not allowed"});
     const listModuleId = req.body;
     for (let i = 0; i < listModuleId.length; i++) {
       await Module.findByIdAndUpdate(listModuleId[i], { position: i + 1 });
